requirements.txt
peewee~=3.19.0
fastapi~=0.128.8
pydantic~=2.8.2
email-validator~=2.2.0
bcrypt~=5.0.0
PyMySQL~=1.1.2
python-dotenv~=1.2.1
pytest~=9.0.2
httpx~=0.28.0
requests~=2.32.5
setuptools~=65.5.0
